refactor(browser): 消除代码重复与性能优化
- 提取 handle_plugin_commands! 宏消除 4 处 Command 处理重复
- 为 plugin 类型实现 From trait，替换手动字段映射为 .into()
- 提取 send_and_wait! 宏消除 channel+timeout 模式重复
- 拆分 BrowserToolOverride::handle 为 6 个独立方法
- 为 maybe_inject_browser_update 添加 5 秒最小检测间隔
- JS Bridge Script 外置为独立 js/bridge.js 文件
- eprintln! 替换为 tracing::debug!/warn!
