+++ crates/tiangong-plugin-browser/Cargo.toml
@@ -14,6 +14,7 @@ tokio = { workspace = true, features = ["rt", "time", "sync"] }
 tauri = { workspace = true, features = ["unstable"] }
 base64-url = "2"
 scru128 = { workspace = true }
+tracing = { workspace = true }
 
 [build-dependencies]
 tauri-plugin = { version = "2", features = ["build"] }
